Feat: 임시 채팅 응답 API 생성, 연결
